refactor : 테스트 성공 후 component에 나눠저 있던 공통 모듈을 common으로 통합했습니다. 모듈들이 통합되면서 의존성도 통합되며 중복은 삭제했습니다.
