+++ application/concert-service/src/main/java/com/fortickets/concertservice/application/client/BookingClient.java
@@ -1,6 +1,6 @@
 package com.fortickets.concertservice.application.client;
 
-import com.fortickets.common.CommonResponse;
+import com.fortickets.common.util.CommonResponse;
 import java.util.List;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.stereotype.Component;
