+++ application/order-service/src/main/java/com/fortickets/orderservice/application/service/PaymentService.java
@@ -1,6 +1,6 @@
 package com.fortickets.orderservice.application.service;
 
-import com.fortickets.common.ErrorCase;
+import com.fortickets.common.util.ErrorCase;
 import com.fortickets.common.exception.GlobalException;
 import com.fortickets.orderservice.application.client.ConcertClient;
 import com.fortickets.orderservice.application.client.UserClient;
