+++ application/order-service/src/main/java/com/fortickets/orderservice/domain/repository/BookingRepository.java
@@ -1,6 +1,6 @@
 package com.fortickets.orderservice.domain.repository;
 
-import com.fortickets.common.BookingStatus;
+import com.fortickets.common.util.BookingStatus;
 import com.fortickets.orderservice.domain.entity.Booking;
 import java.util.List;
 import java.util.Optional;
