+++ application/order-service/src/main/java/com/fortickets/orderservice/infrastructure/config/ComponentConfig.java
@@ -5,9 +5,9 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @ComponentScan(basePackages = {
-        "com.fortickets.exception",
-        "com.fortickets.jpa",
-        "com.fortickets.security",
+    "com.fortickets.common.exception",
+    "com.fortickets.common.jpa",
+    "com.fortickets.common.security",
         "com.fortickets.redis"
 })
 public class ComponentConfig {
