--- application/user-service/src/main/java/com/fortickets/userservice/application/security/UserDetailsServiceImpl.java
@@ -1,28 +1,0 @@
-package com.fortickets.userservice.application.security;
-
-import com.fortickets.userservice.domain.entity.User;
-import com.fortickets.userservice.domain.repository.UserRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    public UserDetailsServiceImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new UsernameNotFoundException("Not Found " + email));
-
-
-        // UserDetailsImpl 객체를 생성할 때 User 객체를 전달합니다.
-        return new UserDetailsImpl(user);
-    }
-}
