+++ application/user-service/src/main/java/com/fortickets/userservice/domain/entity/User.java
@@ -1,8 +1,8 @@
 package com.fortickets.userservice.domain.entity;
 
-import static com.fortickets.jpa.BaseEntity.DELETED_FALSE;
+import static com.fortickets.common.jpa.BaseEntity.DELETED_FALSE;
 
-import com.fortickets.jpa.BaseEntity;
+import com.fortickets.common.jpa.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
