+++ application/user-service/src/main/java/com/fortickets/userservice/infrastructure/config/ComponentConfig.java
@@ -5,9 +5,10 @@ import org.springframework.context.annotation.Configuration;
 
 @Configuration
 @ComponentScan(basePackages = {
-    "com.fortickets.exception",
-    "com.fortickets.jpa",
-    "com.fortickets.security",
+//    "com.fortickets.exception",
+    "com.fortickets.common.exception",
+    "com.fortickets.common.jpa",
+    "com.fortickets.common.security",
     "com.fortickets.redis"
 
 })
