+++ component/common/src/main/java/com/fortickets/common/security/CustomUser.java
@@ -1,4 +1,4 @@
-package com.fortickets.security;
+package com.fortickets.common.security;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
