+++ component/common/src/main/java/com/fortickets/common/security/UseAuth.java
@@ -1,4 +1,4 @@
-package com.fortickets.security;
+package com.fortickets.common.security;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
