+++ component/common/src/main/java/com/fortickets/common/security/UseAuthHandlerMethodArgumentResolver.java
@@ -1,5 +1,7 @@
-package com.fortickets.security;
+package com.fortickets.common.security;
 
+import com.fortickets.common.security.CustomUser;
+import com.fortickets.common.security.UseAuth;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
