+++ component/common/src/main/java/com/fortickets/common/security/WebConfig.java
@@ -1,5 +1,6 @@
-package com.fortickets.security;
+package com.fortickets.common.security;
 
+import com.fortickets.common.security.UseAuthHandlerMethodArgumentResolver;
 import java.util.List;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
