+++ component/security/src/main/java/com/fortickets/security/CustomAuthentication.java
@@ -1,18 +1,18 @@
-package com.fortickets.security;
-
-import java.util.Collection;
-import lombok.Getter;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.GrantedAuthority;
-
-@Getter
-public class CustomAuthentication extends UsernamePasswordAuthenticationToken {
-    private final Long userId;
-    private final String email;
-
-    public CustomAuthentication(Long userId, String email, Collection<? extends GrantedAuthority> authorities) {
-        super(userId, null, authorities);
-        this.userId = userId;
-        this.email = email;
-    }
-}
+//package com.fortickets.security;
+//
+//import java.util.Collection;
+//import lombok.Getter;
+//import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+//import org.springframework.security.core.GrantedAuthority;
+//
+//@Getter
+//public class CustomAuthentication extends UsernamePasswordAuthenticationToken {
+//    private final Long userId;
+//    private final String email;
+//
+//    public CustomAuthentication(Long userId, String email, Collection<? extends GrantedAuthority> authorities) {
+//        super(userId, null, authorities);
+//        this.userId = userId;
+//        this.email = email;
+//    }
+//}
